requirements.txt
anyio==4.8.0
apify_client==1.9.2
apify_shared==1.2.1
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
more-itertools==10.6.0
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.2
anyio==4.8.0
apify_client==1.9.2
apify_shared==1.2.1
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
more-itertools==10.6.0
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
